refactor: apply barrell export
