+++ src/index.js
@@ -4,9 +4,11 @@ import "./index.css";
 import App from "./App";
 import { makeServer } from "./server";
 import { BrowserRouter } from "react-router-dom";
-import { LoginSignupProvider } from "./hooks/context/loginSignup-context";
-import { FilterContextProvider } from "./hooks/context/filter-context";
-import { CartProvider } from "./hooks/context/cart-wishlist-context";
+import {
+  CartProvider,
+  FilterContextProvider,
+  LoginSignupProvider,
+} from "./Context";
 
 // Call make Server
 makeServer();
