+++ src/pages/Authentication/Forgotpassword.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { Link } from "react-router-dom";
-import { Footer, Navbar } from "../../components";
+import {  Navbar } from "../../components";
 import './Authentication.css'
 
 export const Forgotpassword = () => {
