+++ src/pages/Cart/component/index.jsx
@@ -1,0 +1,2 @@
+export { Cardcart } from "./Cardcart";
+export { Cartbill } from "./Cartbill";
